move utils to common module
